Screen after finishing the final level.
That can be done so:
if level == 11:
mängu_etapp = "game_over"

In the game over stage, the program could check if the level is 11. If it is, gameoverwin.png is shown. If it's not, gameover.png is shown.
Reset the level number after the player clicks on one of the buttons (either save or don't save buttons).